+++ app/components/AnalystDashboard/AllDashboard.tsx
@@ -23,7 +23,11 @@ import ClearFilters from 'components/Table/ClearFilters';
 import type { AllDashboardTable_query$key } from '__generated__/AllDashboardTable_query.graphql';
 import { TableCellProps } from '@mui/material';
 import { useFeature } from '@growthbook/growthbook-react';
-import { filterZones, sortZones } from './AssessmentAssignmentTable';
+import {
+  filterZones,
+  sortStatus,
+  sortZones,
+} from './AssessmentAssignmentTable';
 import AdditionalFilters, {
   additionalFilterColumns,
 } from './AdditionalFilters';
@@ -49,15 +53,16 @@ export const filterNumber = (row, id, filterValue) => {
   return numericProperty === Number(filterValue);
 };
 
+// matching the status with `applicant statuses` since cbc only has external status
 const cbcProjectStatusConverter = (status) => {
   if (status === 'Conditionally Approved') {
-    return 'conditionally_approved';
+    return 'applicant_conditionally_approved';
   }
   if (status === 'Reporting Complete') {
     return 'complete';
   }
   if (status === 'Agreement Signed') {
-    return 'approved';
+    return 'applicant_approved';
   }
   if (status === 'Withdrawn') {
     return 'withdrawn';
@@ -116,7 +121,7 @@ const CcbcIdCell = ({ cell }) => {
     <>
       {linkCbc ? (
         <StyledLink
-          href={`/analyst/${isCbcProject ? 'cbc' : 'application'}/${applicationId}`}
+          href={`/analyst/${isCbcProject ? 'cbc' : 'application'}/${applicationId}${isCbcProject ? '' : '/summary'}`}
         >
           {cell.getValue()}
         </StyledLink>
@@ -163,6 +168,12 @@ const AllDashboardTable: React.FC<Props> = ({ query }) => {
     graphql`
       fragment AllDashboardTable_query on Query {
         ...AssignLead_query
+        allApplicationStatusTypes {
+          nodes {
+            name
+            statusOrder
+          }
+        }
         allApplications(
           first: 1000
           filter: { archivedAt: { isNull: true } }
@@ -228,7 +239,8 @@ const AllDashboardTable: React.FC<Props> = ({ query }) => {
     },
     [queryFragment]
   );
-  const { allApplications, allCbcData } = queryFragment;
+  const { allApplications, allCbcData, allApplicationStatusTypes } =
+    queryFragment;
   const isLargeUp = useMediaQuery('(min-width:1007px)');
 
   const [isFirstRender, setIsFirstRender] = useState(true);
@@ -262,6 +274,13 @@ const AllDashboardTable: React.FC<Props> = ({ query }) => {
     projectTitle: 150,
     zones: 91,
   });
+
+  const statusOrderMap = useMemo(() => {
+    return allApplicationStatusTypes?.nodes?.reduce((acc, status) => {
+      acc[status.name] = status.statusOrder;
+      return acc;
+    }, {});
+  }, [allApplicationStatusTypes?.nodes]);
 
   useEffect(() => {
     const sortingSession = cookie.get('mrt_sorting_application');
@@ -394,16 +413,16 @@ const AllDashboardTable: React.FC<Props> = ({ query }) => {
     return [
       ...allApplications.edges.map((application) => ({
         ...application.node,
-        intakeNumber: application.node.ccbcNumber.includes('000074')
+        intakeNumber: application?.node?.ccbcNumber?.includes('000074')
           ? ''
           : application.node.intakeNumber,
         projectId: application.node.ccbcNumber,
         packageNumber: application.node.package,
-        projectTitle:
-          application.node.applicationSowDataByApplicationId?.nodes[0]?.jsonData
-            ?.projectTitle || application.node.projectName,
+        projectTitle: application.node.projectName,
         isCbcProject: false,
         showLink: true,
+        externalStatusOrder: statusOrderMap[application.node.externalStatus],
+        internalStatusOrder: statusOrderMap[application.node.analystStatus],
       })),
       ...(showCbcProjects
         ? allCbcData.edges.map((project) => ({
@@ -412,10 +431,15 @@ const AllDashboardTable: React.FC<Props> = ({ query }) => {
             program: 'CBC',
             zones: [],
             intakeNumber: project.node.jsonData?.intake || 'N/A',
-            projectId: project.node.jsonData.projectNumber,
+            projectId: project.node.projectNumber,
             internalStatus: null,
             externalStatus: project.node.jsonData.projectStatus
               ? cbcProjectStatusConverter(project.node.jsonData.projectStatus)
+              : null,
+            externalStatusOrder: project.node.jsonData.projectStatus
+              ? statusOrderMap[
+                  cbcProjectStatusConverter(project.node.jsonData.projectStatus)
+                ]
               : null,
             packageNumber: null,
             organizationName:
@@ -426,7 +450,13 @@ const AllDashboardTable: React.FC<Props> = ({ query }) => {
           })) ?? []
         : []),
     ];
-  }, [allApplications, allCbcData, showCbcProjects, showCbcProjectsLink]);
+  }, [
+    allApplications.edges,
+    allCbcData.edges,
+    showCbcProjects,
+    showCbcProjectsLink,
+    statusOrderMap,
+  ]);
 
   const columns = useMemo<MRT_ColumnDef<Application>[]>(() => {
     const uniqueIntakeNumbers = [
@@ -501,6 +531,7 @@ const AllDashboardTable: React.FC<Props> = ({ query }) => {
         filterVariant: 'multi-select',
         filterFn: statusFilter,
         filterSelectOptions: analystStatuses,
+        sortingFn: sortStatus,
       },
       {
         accessorKey: 'externalStatus',
@@ -509,6 +540,7 @@ const AllDashboardTable: React.FC<Props> = ({ query }) => {
         filterVariant: 'multi-select',
         filterFn: statusFilter,
         filterSelectOptions: externalStatuses,
+        sortingFn: sortStatus,
       },
       {
         accessorKey: 'projectTitle',