+++ app/tests/pages/analyst/dashboard.test.tsx
@@ -160,6 +160,7 @@ const mockQueryPayload = {
                 conditionalApprovalLetterSent: 'YES',
                 agreementSigned: 'YES',
               },
+              projectNumber: 5555,
             },
           },
           {
@@ -209,6 +210,7 @@ const mockQueryPayload = {
                 conditionalApprovalLetterSent: 'YES',
                 agreementSigned: 'YES',
               },
+              projectNumber: 4444,
             },
           },
           {
@@ -258,6 +260,7 @@ const mockQueryPayload = {
                 conditionalApprovalLetterSent: 'YES',
                 agreementSigned: 'YES',
               },
+              projectNumber: 3333,
             },
           },
         ],
@@ -446,7 +449,7 @@ describe('The index page', () => {
 
     const row = screen.getByText('CCBC-010001');
 
-    expect(row).toHaveAttribute('href', '/analyst/application/1');
+    expect(row).toHaveAttribute('href', '/analyst/application/1/summary');
   });
 
   it('shows the assign lead dropdown when column enabled', async () => {